+++ oap-server/server-storage-plugin/storage-iotdb-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/iotdb/management/IoTDBUITemplateManagementDAO.java
@@ -37,6 +37,7 @@ import org.apache.skywalking.oap.server.library.util.StringUtil;
 import org.apache.skywalking.oap.server.storage.plugin.iotdb.IoTDBClient;
 import org.apache.skywalking.oap.server.storage.plugin.iotdb.IoTDBIndexes;
 import org.apache.skywalking.oap.server.storage.plugin.iotdb.base.IoTDBInsertRequest;
+import org.apache.skywalking.oap.server.storage.plugin.iotdb.utils.IoTDBUtils;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -52,14 +53,14 @@ public class IoTDBUITemplateManagementDAO implements UITemplateManagementDAO {
         }
         StringBuilder query = new StringBuilder();
         query.append("select * from ");
-        query = client.addModelPath(query, UITemplate.INDEX_NAME);
+        IoTDBUtils.addModelPath(client.getStorageGroup(), query, UITemplate.INDEX_NAME);
         Map<String, String> indexAndValueMap = new HashMap<>();
         indexAndValueMap.put(IoTDBIndexes.ID_IDX, id);
-        query = client.addQueryIndexValue(UITemplate.INDEX_NAME, query, indexAndValueMap);
+        IoTDBUtils.addQueryIndexValue(UITemplate.INDEX_NAME, query, indexAndValueMap);
         query.append(" limit 1").append(IoTDBClient.ALIGN_BY_DEVICE);
 
-        List<? super StorageData> storageDataList = client.filterQuery(
-            UITemplate.INDEX_NAME, query.toString(), storageBuilder);
+        List<? super StorageData> storageDataList =
+                client.filterQuery(UITemplate.INDEX_NAME, query.toString(), storageBuilder);
         if (storageDataList.size() > 0) {
             return new DashboardConfiguration().fromEntity((UITemplate) storageDataList.get(0));
         }
@@ -70,20 +71,19 @@ public class IoTDBUITemplateManagementDAO implements UITemplateManagementDAO {
     public List<DashboardConfiguration> getAllTemplates(Boolean includingDisabled) throws IOException {
         StringBuilder query = new StringBuilder();
         query.append("select * from ");
-        query = client.addModelPath(query, UITemplate.INDEX_NAME);
-        query = client.addQueryAsterisk(UITemplate.INDEX_NAME, query);
+        IoTDBUtils.addModelPath(client.getStorageGroup(), query, UITemplate.INDEX_NAME);
+        IoTDBUtils.addQueryAsterisk(UITemplate.INDEX_NAME, query);
         if (!includingDisabled) {
             query.append(" where ").append(UITemplate.DISABLED).append(" = ").append(BooleanUtils.FALSE);
         }
         query.append(IoTDBClient.ALIGN_BY_DEVICE);
 
         List<? super StorageData> storageDataList = client.filterQuery(UITemplate.INDEX_NAME, query.toString(),
-                                                                       storageBuilder
-        );
+                                                                       storageBuilder);
         List<DashboardConfiguration> dashboardConfigurationList = new ArrayList<>(storageDataList.size());
-        storageDataList.forEach(storageData ->
-                                    dashboardConfigurationList.add(
-                                        new DashboardConfiguration().fromEntity((UITemplate) storageData)));
+        storageDataList.forEach(
+                storageData -> dashboardConfigurationList.add(
+                        new DashboardConfiguration().fromEntity((UITemplate) storageData)));
         return dashboardConfigurationList;
     }
 
@@ -91,9 +91,9 @@ public class IoTDBUITemplateManagementDAO implements UITemplateManagementDAO {
     public TemplateChangeStatus addTemplate(DashboardSetting setting) throws IOException {
         final UITemplate uiTemplate = setting.toEntity();
 
-        IoTDBInsertRequest request = new IoTDBInsertRequest(UITemplate.INDEX_NAME, UI_TEMPLATE_TIMESTAMP,
-                                                            uiTemplate, storageBuilder
-        );
+        IoTDBInsertRequest request =
+                IoTDBInsertRequest.buildRequest(UITemplate.INDEX_NAME, UI_TEMPLATE_TIMESTAMP,
+                                                uiTemplate, storageBuilder);
         client.write(request);
         return TemplateChangeStatus.builder().status(true).id(setting.getId()).build();
     }
@@ -104,11 +104,11 @@ public class IoTDBUITemplateManagementDAO implements UITemplateManagementDAO {
 
         StringBuilder query = new StringBuilder();
         query.append("select * from ");
-        query = client.addModelPath(query, UITemplate.INDEX_NAME);
-        query.append(IoTDBClient.DOT).append(client.indexValue2LayerName(uiTemplate.id()))
+        IoTDBUtils.addModelPath(client.getStorageGroup(), query, UITemplate.INDEX_NAME);
+        query.append(IoTDBClient.DOT).append(IoTDBUtils.indexValue2LayerName(uiTemplate.id()))
              .append(IoTDBClient.ALIGN_BY_DEVICE);
         List<? super StorageData> queryResult = client.filterQuery(
-            UITemplate.INDEX_NAME, query.toString(), storageBuilder);
+                UITemplate.INDEX_NAME, query.toString(), storageBuilder);
         if (queryResult.size() == 0) {
             return TemplateChangeStatus.builder()
                                        .status(false)
@@ -116,9 +116,9 @@ public class IoTDBUITemplateManagementDAO implements UITemplateManagementDAO {
                                        .message("Can't find the template")
                                        .build();
         } else {
-            IoTDBInsertRequest request = new IoTDBInsertRequest(UITemplate.INDEX_NAME, UI_TEMPLATE_TIMESTAMP,
-                                                                uiTemplate, storageBuilder
-            );
+            IoTDBInsertRequest request =
+                    IoTDBInsertRequest.buildRequest(UITemplate.INDEX_NAME, UI_TEMPLATE_TIMESTAMP,
+                                                    uiTemplate, storageBuilder);
             client.write(request);
             return TemplateChangeStatus.builder().status(true).id(setting.getId()).build();
         }
@@ -128,20 +128,20 @@ public class IoTDBUITemplateManagementDAO implements UITemplateManagementDAO {
     public TemplateChangeStatus disableTemplate(String id) throws IOException {
         StringBuilder query = new StringBuilder();
         query.append("select * from ");
-        query = client.addModelPath(query, UITemplate.INDEX_NAME);
-        query.append(IoTDBClient.DOT).append(client.indexValue2LayerName(id))
+        IoTDBUtils.addModelPath(client.getStorageGroup(), query, UITemplate.INDEX_NAME);
+        query.append(IoTDBClient.DOT).append(IoTDBUtils.indexValue2LayerName(id))
              .append(IoTDBClient.ALIGN_BY_DEVICE);
 
         List<? super StorageData> queryResult = client.filterQuery(
-            UITemplate.INDEX_NAME, query.toString(), storageBuilder);
+                UITemplate.INDEX_NAME, query.toString(), storageBuilder);
         if (queryResult.size() == 0) {
             return TemplateChangeStatus.builder().status(false).id(id).message("Can't find the template").build();
         } else {
             final UITemplate uiTemplate = (UITemplate) queryResult.get(0);
             uiTemplate.setDisabled(BooleanUtils.TRUE);
-            IoTDBInsertRequest request = new IoTDBInsertRequest(UITemplate.INDEX_NAME, UI_TEMPLATE_TIMESTAMP,
-                                                                uiTemplate, storageBuilder
-            );
+            IoTDBInsertRequest request =
+                    IoTDBInsertRequest.buildRequest(UITemplate.INDEX_NAME, UI_TEMPLATE_TIMESTAMP,
+                                                    uiTemplate, storageBuilder);
             client.write(request);
             return TemplateChangeStatus.builder().status(true).id(id).build();
         }