+++ oap-server/server-storage-plugin/storage-iotdb-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/iotdb/query/IoTDBTopologyQueryDAO.java
@@ -18,6 +18,7 @@
 
 package org.apache.skywalking.oap.server.storage.plugin.iotdb.query;
 
+import com.google.common.base.Splitter;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
@@ -39,6 +40,7 @@ import org.apache.skywalking.oap.server.core.query.type.Call;
 import org.apache.skywalking.oap.server.core.source.DetectPoint;
 import org.apache.skywalking.oap.server.core.storage.query.ITopologyQueryDAO;
 import org.apache.skywalking.oap.server.storage.plugin.iotdb.IoTDBClient;
+import org.apache.skywalking.oap.server.storage.plugin.iotdb.utils.IoTDBUtils;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -47,7 +49,8 @@ public class IoTDBTopologyQueryDAO implements ITopologyQueryDAO {
 
     @Override
     public List<Call.CallDetail> loadServiceRelationsDetectedAtServerSide(long startTB, long endTB,
-                                                                          List<String> serviceIds) throws IOException {
+                                                                          List<String> serviceIds)
+            throws IOException {
         return loadServiceCalls(
                 ServiceRelationServerSideMetrics.INDEX_NAME, startTB, endTB,
                 ServiceRelationServerSideMetrics.SOURCE_SERVICE_ID,
@@ -57,7 +60,8 @@ public class IoTDBTopologyQueryDAO implements ITopologyQueryDAO {
 
     @Override
     public List<Call.CallDetail> loadServiceRelationDetectedAtClientSide(long startTB, long endTB,
-                                                                         List<String> serviceIds) throws IOException {
+                                                                         List<String> serviceIds)
+            throws IOException {
         return loadServiceCalls(
                 ServiceRelationClientSideMetrics.INDEX_NAME, startTB, endTB,
                 ServiceRelationClientSideMetrics.SOURCE_SERVICE_ID,
@@ -66,7 +70,8 @@ public class IoTDBTopologyQueryDAO implements ITopologyQueryDAO {
     }
 
     @Override
-    public List<Call.CallDetail> loadServiceRelationsDetectedAtServerSide(long startTB, long endTB) throws IOException {
+    public List<Call.CallDetail> loadServiceRelationsDetectedAtServerSide(long startTB, long endTB)
+            throws IOException {
         return loadServiceCalls(
                 ServiceRelationServerSideMetrics.INDEX_NAME, startTB, endTB,
                 ServiceRelationServerSideMetrics.SOURCE_SERVICE_ID,
@@ -75,7 +80,8 @@ public class IoTDBTopologyQueryDAO implements ITopologyQueryDAO {
     }
 
     @Override
-    public List<Call.CallDetail> loadServiceRelationDetectedAtClientSide(long startTB, long endTB) throws IOException {
+    public List<Call.CallDetail> loadServiceRelationDetectedAtClientSide(long startTB, long endTB)
+            throws IOException {
         return loadServiceCalls(
                 ServiceRelationClientSideMetrics.INDEX_NAME, startTB, endTB,
                 ServiceRelationClientSideMetrics.SOURCE_SERVICE_ID,
@@ -86,7 +92,8 @@ public class IoTDBTopologyQueryDAO implements ITopologyQueryDAO {
     @Override
     public List<Call.CallDetail> loadInstanceRelationDetectedAtServerSide(String clientServiceId,
                                                                           String serverServiceId,
-                                                                          long startTB, long endTB) throws IOException {
+                                                                          long startTB, long endTB)
+            throws IOException {
         return loadServiceInstanceCalls(
                 ServiceInstanceRelationServerSideMetrics.INDEX_NAME, startTB, endTB,
                 ServiceInstanceRelationServerSideMetrics.SOURCE_SERVICE_ID,
@@ -97,7 +104,8 @@ public class IoTDBTopologyQueryDAO implements ITopologyQueryDAO {
     @Override
     public List<Call.CallDetail> loadInstanceRelationDetectedAtClientSide(String clientServiceId,
                                                                           String serverServiceId,
-                                                                          long startTB, long endTB) throws IOException {
+                                                                          long startTB, long endTB)
+            throws IOException {
         return loadServiceInstanceCalls(
                 ServiceInstanceRelationClientSideMetrics.INDEX_NAME, startTB, endTB,
                 ServiceInstanceRelationClientSideMetrics.SOURCE_SERVICE_ID,
@@ -106,7 +114,8 @@ public class IoTDBTopologyQueryDAO implements ITopologyQueryDAO {
     }
 
     @Override
-    public List<Call.CallDetail> loadEndpointRelation(long startTB, long endTB, String destEndpointId) throws IOException {
+    public List<Call.CallDetail> loadEndpointRelation(long startTB, long endTB, String destEndpointId)
+            throws IOException {
         List<Call.CallDetail> calls = loadEndpointFromSide(
                 EndpointRelationServerSideMetrics.INDEX_NAME, startTB, endTB,
                 EndpointRelationServerSideMetrics.SOURCE_ENDPOINT,
@@ -122,20 +131,20 @@ public class IoTDBTopologyQueryDAO implements ITopologyQueryDAO {
 
     private List<Call.CallDetail> loadServiceCalls(String tableName, long startTB, long endTB,
                                                    String sourceCName, String destCName,
-                                                   List<String> serviceIds, DetectPoint detectPoint) throws IOException {
+                                                   List<String> serviceIds, DetectPoint detectPoint)
+            throws IOException {
         // This method don't use "group by" like other storage plugin.
         StringBuilder query = new StringBuilder();
         query.append("select ").append(ServiceRelationServerSideMetrics.COMPONENT_ID).append(" from ");
-        query = client.addModelPath(query, tableName);
-        query = client.addQueryAsterisk(tableName, query);
+        IoTDBUtils.addModelPath(client.getStorageGroup(), query, tableName);
+        IoTDBUtils.addQueryAsterisk(tableName, query);
         query.append(" where ").append(IoTDBClient.TIME).append(" >= ").append(TimeBucket.getTimestamp(startTB))
-                .append(" and ").append(IoTDBClient.TIME).append(" <= ").append(TimeBucket.getTimestamp(endTB));
+             .append(" and ").append(IoTDBClient.TIME).append(" <= ").append(TimeBucket.getTimestamp(endTB));
         if (serviceIds.size() > 0) {
             query.append(" and (");
             for (int i = 0; i < serviceIds.size(); i++) {
-                query.append(sourceCName).append(" = \"").append(serviceIds.get(i))
-                        .append("\" or ")
-                        .append(destCName).append(" = \"").append(serviceIds.get(i)).append("\"");
+                query.append(sourceCName).append(" = \"").append(serviceIds.get(i)).append("\" or ")
+                     .append(destCName).append(" = \"").append(serviceIds.get(i)).append("\"");
                 if (i != serviceIds.size() - 1) {
                     query.append(" or ");
                 }
@@ -157,8 +166,9 @@ public class IoTDBTopologyQueryDAO implements ITopologyQueryDAO {
                 RowRecord rowRecord = wrapper.next();
                 List<Field> fields = rowRecord.getFields();
                 Call.CallDetail call = new Call.CallDetail();
-                String[] layerNames = fields.get(0).getStringValue().split("\\" + IoTDBClient.DOT + "\"");
-                String entityId = client.layerName2IndexValue(layerNames[2]);
+                List<String> layerNames = Splitter.on(IoTDBClient.DOT + "\"")
+                                                  .splitToList(fields.get(0).getStringValue());
+                String entityId = IoTDBUtils.layerName2IndexValue(layerNames.get(2));
                 final int componentId = fields.get(1).getIntV();
                 call.buildFromServiceRelation(entityId, componentId, detectPoint);
                 calls.add(call);
@@ -180,15 +190,15 @@ public class IoTDBTopologyQueryDAO implements ITopologyQueryDAO {
         // This method don't use "group by" like other storage plugin.
         StringBuilder query = new StringBuilder();
         query.append("select ").append(ServiceInstanceRelationServerSideMetrics.COMPONENT_ID).append(" from ");
-        query = client.addModelPath(query, tableName);
-        query = client.addQueryAsterisk(tableName, query);
+        IoTDBUtils.addModelPath(client.getStorageGroup(), query, tableName);
+        IoTDBUtils.addQueryAsterisk(tableName, query);
         query.append(" where ").append(IoTDBClient.TIME).append(" >= ").append(TimeBucket.getTimestamp(startTB))
-                .append(" and ").append(IoTDBClient.TIME).append(" <= ").append(TimeBucket.getTimestamp(endTB));
+             .append(" and ").append(IoTDBClient.TIME).append(" <= ").append(TimeBucket.getTimestamp(endTB));
         query.append(" and ((").append(sourceCName).append(" = \"").append(sourceServiceId).append("\"")
-                .append(" and ").append(descCName).append(" = \"").append(destServiceId).append("\"")
-                .append(") or (").append(sourceCName).append(" = \"").append(destServiceId).append("\"")
-                .append(" and ").append(descCName).append(" = \"").append(sourceServiceId).append(" \"))")
-                .append(IoTDBClient.ALIGN_BY_DEVICE);
+             .append(" and ").append(descCName).append(" = \"").append(destServiceId).append("\"")
+             .append(") or (").append(sourceCName).append(" = \"").append(destServiceId).append("\"")
+             .append(" and ").append(descCName).append(" = \"").append(sourceServiceId).append(" \"))")
+             .append(IoTDBClient.ALIGN_BY_DEVICE);
 
         SessionPool sessionPool = client.getSessionPool();
         SessionDataSetWrapper wrapper = null;
@@ -202,8 +212,9 @@ public class IoTDBTopologyQueryDAO implements ITopologyQueryDAO {
                 RowRecord rowRecord = wrapper.next();
                 List<Field> fields = rowRecord.getFields();
                 Call.CallDetail call = new Call.CallDetail();
-                String[] layerNames = fields.get(0).getStringValue().split("\\" + IoTDBClient.DOT + "\"");
-                String entityId = client.layerName2IndexValue(layerNames[2]);
+                List<String> layerNames = Splitter.on(IoTDBClient.DOT + "\"")
+                                                  .splitToList(fields.get(0).getStringValue());
+                String entityId = IoTDBUtils.layerName2IndexValue(layerNames.get(2));
                 final int componentId = fields.get(1).getIntV();
                 call.buildFromInstanceRelation(entityId, componentId, detectPoint);
                 calls.add(call);
@@ -224,12 +235,12 @@ public class IoTDBTopologyQueryDAO implements ITopologyQueryDAO {
         // This method don't use "group by" like other storage plugin.
         StringBuilder query = new StringBuilder();
         query.append("select * from ");
-        query = client.addModelPath(query, tableName);
-        query = client.addQueryAsterisk(tableName, query);
+        IoTDBUtils.addModelPath(client.getStorageGroup(), query, tableName);
+        IoTDBUtils.addQueryAsterisk(tableName, query);
         query.append(" where ").append(IoTDBClient.TIME).append(" >= ").append(TimeBucket.getTimestamp(startTB))
-                .append(" and ").append(IoTDBClient.TIME).append(" <= ").append(TimeBucket.getTimestamp(endTB));
+             .append(" and ").append(IoTDBClient.TIME).append(" <= ").append(TimeBucket.getTimestamp(endTB));
         query.append(" and ").append(isSourceId ? sourceCName : destCName).append(" = \"").append(id).append("\"")
-                .append(IoTDBClient.ALIGN_BY_DEVICE);
+             .append(IoTDBClient.ALIGN_BY_DEVICE);
 
         SessionPool sessionPool = client.getSessionPool();
         SessionDataSetWrapper wrapper = null;
@@ -243,8 +254,9 @@ public class IoTDBTopologyQueryDAO implements ITopologyQueryDAO {
                 RowRecord rowRecord = wrapper.next();
                 List<Field> fields = rowRecord.getFields();
                 Call.CallDetail call = new Call.CallDetail();
-                String[] layerNames = fields.get(0).getStringValue().split("\\" + IoTDBClient.DOT + "\"");
-                String entityId = client.layerName2IndexValue(layerNames[2]);
+                List<String> layerNames = Splitter.on(IoTDBClient.DOT + "\"")
+                                                  .splitToList(fields.get(0).getStringValue());
+                String entityId = IoTDBUtils.layerName2IndexValue(layerNames.get(2));
                 call.buildFromEndpointRelation(entityId, DetectPoint.SERVER);
                 calls.add(call);
             }